moved tests to tests/calculators/risk/
